requirements.txt
amqp==5.1.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.5
Django==4.1.8
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
exceptiongroup==1.1.1
iniconfig==2.0.0
kombu==5.2.4
packaging==23.1
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
pymongo==4.3.3
pytest==7.3.1
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2023.3
six==1.16.0
sqlparse==0.2.4
tomli==2.0.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6